Move get_proposal to the base document mixin
Some views are shared between mails and documents (sinde Mail is a
Document as well).
Since the mail could be linked to a proposal as well, the get_proposal
method should be available for the email as well.
Therefore it needs to be moved to the BaseMixin.
